# Optimize FullProjectSnapshotHandle Deserialization