… src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Performance/ProjectSystem/project.razor.json