This was in a fairly hot path and could be optimized pretty easily by leveraging the Reader instead of converting to JObject.

![image](https://user-images.githubusercontent.com/14852843/87706951-309d8a80-c755-11ea-80c0-708cdf553a45.png)

We already tested the round trip serialization validity [here](https://github.com/dotnet/aspnetcore-tooling/blob/master/src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Common.Test/SerializationTest.cs) so I didn't add any new tests.

|                                                              Method |     Mean |     Error |    StdDev |  Op/s |   Gen 0 |   Gen 1 | Allocated |
|-------------------------------------------------------------------- |---------:|----------:|----------:|------:|--------:|--------:|----------:|
|   'Razor FullProjectSnapshotHandle Roundtrip JObject Serialization' | 7.590 ms | 0.0450 ms | 0.0376 ms | 131.8 | 62.5000 | 31.2500 |   6.81 MB |
| 'Razor ProjectSnapshotHandle Roundtrip JsonConverter Serialization' | 3.185 ms | 0.0151 ms | 0.0134 ms | 314.0 | 35.1563 | 15.6250 |   2.81 MB |

Fixes: https://github.com/dotnet/aspnetcore/issues/24028