@TanayParikh 2020-07-17
Added benchmarks - 66% lower memory usage & 58% less CPU time